Create separate About and Partners pages
