+++ src/pages/partners.module.scss
@@ -1,0 +1,21 @@
+@import "../_assets/css/abstracts/1-abstracts";
+/*
+ *
+ * Partners
+ *
+ */
+
+.partnersTitle {
+  margin-top: var(--sp-loosest);
+  margin-bottom: var(--sp-loose);
+}
+
+.logoGroup {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(auto, 160px));
+  gap: var(--sp-loosest);
+}
+
+.logoItem {
+  align-self: center;
+}
